murmur-engine/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
openai==1.14.3
pydantic==2.6.4
SQLAlchemy==2.0.35
